[VxTwit] Add `ddinstagram.com` replacements (#656)
* [VxTwit] Add `ddinstagram.com` replacements

This commit adds event listeners to replace instagram.com links
with ddinstagram.com. It is using the same toggle as VxTwit.

Since instagram.com embeds don't have videos or photos, we will need to
replace all embeds. As such, the `valid` method was modified to not
check for videos in embeds: this is deferred to a later check for the
VxTwitter case.

This commit doesn't change the cog name; this will be done in a follow
up.

* [VxTwit] Use Enum for SocialMedia

* [VxTwit] Remove unnecessary debug statement

* [VxTwit] Sort the imports

* [VxTwit] Move INSTA_REGEX_MATCH to constants.py

* [VxTwit] Refactor urls_to_string

Co-authored-by: Dat Quach <quachtridat@yahoo.com.vn>

* [VxTwit] Fix docstrings

* [VxTwit] Compile regex pattern

Co-authored-by: Dat Quach <quachtridat@yahoo.com.vn>

* [VxTwit] Add on_message_edit listener for instagram