[Birthday] Add test coverage
This is an extensively large commit that adds, readjusts and
consolidates tests covering most functions and methods in
`commandsCore.py`, `converters.py`, and `core.py`.

Minor adjustments are made to:
- `commandsCore.py` so that only `datetime.date` needs to be imported.
- `converters.py` to rely on `datetime.date` only.

The new file `conftest.py` contains multiple Pytest fixtures to be used
by a large number of tests. Some fixtures here are made to patch
specific attributes of Red and the Birthday cog to simulate the cog as
close to production as possible. Locale setup is also done as an
"auto-use" fixture in this file, which reduces the locale boilerplate
code in other test files, as this cog interacts with date objects
actively.

`testBirthday.py` is deleted in this commit. It is superseded by
`testCore.py`.

Some adjustments are done to `testConverters.py` to make the test
compact.

Note that there is an extensive use of the type hint pattern
`typing.Union[mock.Mock, Type]`. This is done so that Python language
servers can apply syntax highlighting instructions based on the `Type`,
while the objects in test stay as `mock.Mock` instances.