[CI] Collect only testful cogs in cog unit test workflow
This commit relies on Pytest's test collection logic to extract the
names of the cogs with tests. Tox is used to keep the
`red-discordbot[test]` base dependency central in one INI file. Cogs
with no tests will not be included in the `cogs` output of the
`discover` job, meaning they will not get recognized by the `test` job.